Modification of the architecture of the API import part
